Seperate out AddCustomExercise
