+++ fittrak-client/src/components/exercises/ExerciseActions.vue
@@ -22,59 +22,7 @@
         Add Exercises
       </v-btn>
 
-      <v-dialog 
-        v-model="dialog" 
-        fullscreen 
-        hide-overlay 
-        transition="dialog-bottom-transition">
-        <v-btn 
-          slot="activator" 
-          color="primary" 
-          dark>Add Custom</v-btn>
-        <v-card>
-          <v-toolbar 
-            dark 
-            color="primary">
-            <v-btn 
-              icon 
-              dark 
-              @click.native="dialog = false">
-              <v-icon>close</v-icon>
-            </v-btn>
-            <v-toolbar-title>Add Custom Exercise</v-toolbar-title>
-            <v-spacer/>
-            <v-toolbar-items>
-              <v-btn 
-                dark 
-                flat 
-                @click.native="addCustomExercise">Save</v-btn>
-            </v-toolbar-items>
-          </v-toolbar>
-
-          <v-card-text>
-            <v-layout wrap>
-              <v-flex xs12>
-                <v-text-field 
-                  placeholder="Exercise name" 
-                  v-model="customExerciseName" />
-              </v-flex>
-
-              <v-flex xs12>
-                <v-autocomplete
-                  v-model="customMuscleGroupName"
-                  :items="muscleGroups"
-                  placeholder="Select muscle group"
-                  browser-autocomplete
-                  clearable
-                />
-              </v-flex>
-
-            </v-layout>
-          </v-card-text>
-
-        </v-card>
-      </v-dialog>
-
+      <AddCustomExercise :workout="workout" />
     </v-form>
   </v-flex>
 </template>
@@ -82,90 +30,26 @@
 <script>
 import { queries, mutations } from "@/graphql";
 
+import AddCustomExercise from "@/components/exercises/AddCustomExercise";
+
 import { showSnackbar } from "@/helpers";
 
 export default {
   name: "AddExerciseForm",
 
   data: () => ({
-    dialog: false,
     newExercises: [],
-    exerciseTypes: [],
-    muscleGroups: [],
-    customMuscleGroupName: "",
-    customExerciseName: ""
+    exerciseTypes: []
   }),
 
   apollo: {
     exerciseTypes: {
       query: queries.exerciseTypesQuery,
       update: data => data.exerciseTypes.map(exerciseType => exerciseType.name)
-    },
-
-    muscleGroups: {
-      query: queries.muscleGroupsQuery,
-      update: data => data.muscleGroups.map(group => group.name)
     }
   },
 
   methods: {
-    addCustomExercise() {
-      const { customMuscleGroupName, customExerciseName } = this;
-      const { workout } = this.$props;
-
-      if (!customMuscleGroupName.length || !customExerciseName.length) {
-        return;
-      }
-
-      this.$apollo
-        .mutate({
-          mutation: mutations.addCustomExerciseMutation,
-
-          variables: {
-            workoutId: workout.id,
-            exerciseName: customExerciseName,
-            muscleGroupName: customMuscleGroupName
-          },
-
-          update(store, { data }) {
-            const customExercise = data.addCustomExercise.exercise;
-
-            const result = store.readQuery({
-              query: queries.exercisesQuery,
-              variables: {
-                workoutId: workout.id
-              }
-            });
-
-            result.exercises = [customExercise, ...result.exercises];
-
-            store.writeQuery({
-              query: queries.exercisesQuery,
-              variables: {
-                workoutId: workout.id
-              },
-              data: result
-            });
-          }
-        })
-        .then(() => {
-          showSnackbar(
-            "success",
-            `Added exercise "${this.customExerciseName}".`
-          );
-
-          this.dialog = false;
-          this.customMuscleGroupName = "";
-          this.customExerciseName = "";
-        })
-        .catch(() => {
-          showSnackbar(
-            "error",
-            "Could not add custom exercise. Support has been notified."
-          );
-        });
-    },
-
     addExercises() {
       const workout = this.workout;
 
@@ -216,6 +100,10 @@ export default {
           );
         });
     }
+  },
+
+  components: {
+    AddCustomExercise
   },
 
   props: {
